Release Version 22.03.02 - Various bugfixes.
* Fixed bug in AssemblyInfo.cs: missing '+'.
* Fixed bug corrected LONG registers not assigned correctly in Cog - DIR, OUT on Cog.cs.
* Fixed bug on LogicAnalog.Draw(): was comparing wrong time to activate.
* Fixed bug in maximum frequency of Cog Counters, to 128Mhz instead of wrong limit of 120Mhz, on PLL.cs.
* Fixed bug on LogicView.cs: save TimeUnitSelected with the other properties in logic view.
* Fixed bug of Form name not updated when open another binary file in Emulator.cs.
* Fixed hidden vertical scroll bar on Logic View: LogicView.Designer.cs.
* Fixed bug of not remembering last binary open from Settings in Emulator.cs and GearDesktop.
* Method PluginEditor.OpenButton_Click(.) refactored on using Open File Dialog.
* Updated GEAR Developers.

+++ Gear/EmulationCore/Cog.cs
@@ -179,14 +179,17 @@ namespace Gear.EmulationCore
         }
 
         /// @brief Property to return complete OUT pins (P0..P63)
-        /// Analyze all sources of pin changes in the cog: OUTA, OUTB, the two counters 
-        /// and the video generator.
+        /// @details Analyze all sources of pin changes in the cog: <c>OUTA</c>,
+        /// <c>OUTB</c>, the two counters and the video generator.
+        /// @version v22.03.02 - Bugfix corrected LONG registers not assigned
+        /// correctly in %Cog - <c>DIR</c>, <c>OUT</c>.
         public ulong OUT
         {
             get
             {
                 return Memory[(int)Assembly.RegisterAddress.OUTA] |
-                    (Memory[(int)Assembly.RegisterAddress.OUTB] << 32) |
+                    //Bugfix OUTB must be shifted on 64 bits long
+                    ((ulong)Memory[(int)Assembly.RegisterAddress.OUTB] << 32) |
                     FreqA.Output |
                     FreqB.Output |
                     Video.Output;
@@ -247,13 +250,16 @@ namespace Gear.EmulationCore
             }
         }
 
-        /// @brief Property to return complete DIR pins.
+        /// @brief Property to return complete register of <c>DIR</c> pins
+        /// @version v22.03.02 - Bugfix corrected LONG registers not assigned
+        /// correctly in %Cog - <c>DIR</c>, <c>OUT</c>.
         public ulong DIR
         {
             get
             {
-                return (Memory[(int)Assembly.RegisterAddress.DIRB] << 32) |
-                    Memory[(int)Assembly.RegisterAddress.DIRA];
+                //Bugfix DIRB must be shifted on 64 bits long
+                return ((ulong)Memory[(int)Assembly.RegisterAddress.DIRB] << 32) |
+                       Memory[(int)Assembly.RegisterAddress.DIRA];
             }
         }
